refactor(openapi-codegen): update jest config file, move test files to another folder
