+++ jest.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-    clearMocks: true,
-    coverageDirectory: 'coverage',
-    preset: 'ts-jest',
-    testEnvironment: 'node',
+    roots: ['<rootDir>/tests'],
+    transform: {
+        '^.+\\.tsx?$': 'ts-jest',
+    },
     coverageThreshold: {
         global: {
             branches: 75,
